m3_get_yll_o3_ecoloss

Produce Economic damages associated with YLLs attributable to O3 (M6M).The economic valuation takes as a base value the Value of Statistical Life Year (VSLY) for EU from Schlander et al (2017) and expands the value to other regions based on the“unit value transfer approach” which adjusts the VSLY according to their GDP and GDP growth rates. YLL-to-Mortalities ratios are based on TM5-FASST calculations. Premature mortalities are based on the integrated exposure-response functions (IER) from Burnett et al (2014), consistent with the GBD 2016 study.

Usage

m3_get_yll_o3_ecoloss(
  db_path = NULL,
  query_path = "./inst/extdata",
  db_name = NULL,
  prj_name,
  rdata_name = NULL,
  scen_name,
  queries = "queries_rfasst.xml",
  final_db_year = 2100,
  ssp = "SSP2",
  saveOutput = T,
  map = F,
  anim = T,
  mort_param = "mort_o3_gbd2016_med",
  recompute = F
)

Source

Schlander, M., Schaefer, R. and Schwarz, O., 2017. Empirical studies on the economic value of a Statistical Life Year (VSLY) in Europe: what do they tell us?. Value in Health, 20(9), p.A666.

Arguments

db_path

Path to the GCAM database

query_path

Path to the query file

db_name

Name of the GCAM database

prj_name

Name of the rgcam project. This can be an existing project, or, if not, this will be the name

rdata_name

Name of the RData file. It must contain the queries in a list

scen_name

Name of the GCAM scenario to be processed

queries

Name of the GCAM query file. The file by default includes the queries required to run rfasst

final_db_year

Final year in the GCAM database (this allows to process databases with user-defined "stop periods")

ssp

Set the ssp narrative associated to the GCAM scenario. c("SSP1","SSP2","SSP3","SSP4","SSP5"). By default is SSP2

saveOutput

Writes the emission files.By default=T

map

Produce the maps. By default=F

anim

If set to T, produces multi-year animations. By default=T

mort_param

Select the health function (GBD 2016 or Jerret et al 2009) and the Low/Med/High RR. By default = mort_o3_gbd2016_med

recompute

If set to T, recomputes the function output. Otherwise, if the output was already computed once, it uses that value and avoids repeating computations. By default=F

Value

Economic damages associated with YLLs attributable to O3 (M6M) exposure for each TM5-FASST regions for all years (Thous$2015). The list of countries that form each region and the full name of the region can be found in Table S2.2 in the TM5-FASST documentation paper: Van Dingenen, R., Dentener, F., Crippa, M., Leitao, J., Marmer, E., Rao, S., Solazzo, E. and Valentini, L., 2018. TM5-FASST: a global atmospheric source-receptor model for rapid impact analysis of emission changes on air quality and short-lived climate pollutants. Atmospheric Chemistry and Physics, 18(21), pp.16173-16211.