m3_get_yll_pm25

Produce YLLs attributable to PM2.5 exposure. YLL-to-Mortalities ratios are based on TM5-FASST calculations. Premature mortalities are based on the integrated exposure-response functions (IER) from Burnett et al (2014), consistent with the GBD 2016 study.

m3_get_yll_pm25(
  db_path = NULL,
  query_path = "./inst/extdata",
  db_name = NULL,
  prj_name,
  prj = NULL,
  scen_name,
  queries = "queries_rfasst.xml",
  final_db_year = 2100,
  mort_param = "GBD",
  ssp = "SSP2",
  saveOutput = T,
  map = F,
  anim = T,
  recompute = F,
  gcam_eur = F
)

Arguments

db_path

Path to the GCAM database

query_path

Path to the query file

db_name

Name of the GCAM database

prj_name

Name of the rgcam project. This can be an existing project, or, if not, this will be the name

prj

rgcam loaded project

scen_name

Vector names of the GCAM scenarios to be processed

queries

Name of the GCAM query file. The file by default includes the queries required to run rfasst

final_db_year

Final year in the GCAM database (this allows to process databases with user-defined "stop periods")

mort_param

Method to compute the premature mortality estimates used in the plotting: GBD, GEMM, or FUSION

ssp

Set the ssp narrative associated to the GCAM scenario. c("SSP1","SSP2","SSP3","SSP4","SSP5"). By default is SSP2

saveOutput

Writes the emission files.By default=T

map

Produce the maps. By default=F

anim

If set to T, produces multi-year animations. By default=T

recompute

If set to T, recomputes the function output. Otherwise, if the output was already computed once, it uses that value and avoids repeating computations. By default=F

gcam_eur

If set to T, considers the GCAM-Europe regions. By default=F

Value

YLLs attributable to PM2.5 exposure for each TM5-FASST regions for all years (# YLLs). The list of countries that form each region and the full name of the region can be found in Table S2.2 in the TM5-FASST documentation paper: Van Dingenen, R., Dentener, F., Crippa, M., Leitao, J., Marmer, E., Rao, S., Solazzo, E. and Valentini, L., 2018. TM5-FASST: a global atmospheric source-receptor model for rapid impact analysis of emission changes on air quality and short-lived climate pollutants. Atmospheric Chemistry and Physics, 18(21), pp.16173-16211.